[DESKTOP] Refactoring single form fields
Some fields were duplicated or could be redone into a single field.
