+++ projects/desktop/src/components/common/SingleForm/atoms/checkbox/CheckboxCheckmark.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { colors } from '../../../../constants'
+import { colors } from '../../../../../constants'
 
 // Custom checkbox checkmark style
 export const CheckboxCheckmark = styled.span`
