+++ projects/desktop/src/components/common/SingleForm/atoms/checkbox/index.ts
@@ -1,0 +1,3 @@
+export { CheckboxContainer } from './CheckboxContainer'
+export { CheckboxInput } from './CheckboxInput'
+export { CheckboxCheckmark } from './CheckboxCheckmark'
