+++ projects/desktop/src/components/common/SingleForm/atoms/input/InputBorder.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { colors } from '../../../../constants'
+import { colors } from '../../../../../constants'
 
 /**
  * The input's border style
