+++ projects/desktop/src/components/common/SingleForm/fields/CheckboxField.tsx
@@ -2,31 +2,32 @@ import React from 'react'
 import { Field } from 'react-final-form'
 import { useTranslation, UseTranslationResponse } from 'react-i18next'
 
-import { CheckboxContainer, CheckboxInput, CheckboxCheckmark } from '../../common/singleForm/index'
+import { CheckboxContainer, CheckboxInput, CheckboxCheckmark } from '../atoms/checkbox'
 
+/**
+ * Checkbox input for single forms
+ */
 interface IOwnProps {
-    invalid: boolean,
+    name: string,
+    textTranslationKey: string,
+    invalid?: boolean,
 }
 
 type IProps = IOwnProps
 
-/**
- * Checkbox field of the register form
- */
-// TODO: can be refactored into a generic component
-export const CheckboxField = ({invalid}: IProps): JSX.Element => {
+export const CheckboxField = ({name, textTranslationKey, invalid }: IProps): JSX.Element => {
     const {t}: UseTranslationResponse = useTranslation()
 
     return (
         <>
             <Field
-                name='termsOfUse'
+                name={name}
                 component='input'
                 type='checkbox'
             >
                 {({input}) => (
                     <CheckboxContainer pose={invalid ? 'invalid' : 'valid'}>
-                        {t('register.form.termsOfUse')}
+                        {t(textTranslationKey)}
                         <CheckboxInput {...input} />
                         <CheckboxCheckmark />
                     </CheckboxContainer>
