+++ projects/desktop/src/components/common/SingleForm/fields/EmailField.tsx
@@ -2,13 +2,12 @@ import React from 'react'
 import { Field } from 'react-final-form'
 import { useTranslation, UseTranslationResponse } from 'react-i18next'
 
-import { Input } from './'
-import { composeValidators, required, validateEmail } from '../../../constants/validators'
+import { Input } from '../atoms/input/Input'
+import { composeValidators, required, validateEmail } from '../../../../constants/validators'
 
 /**
- * Email field of the login form
+ * Email field for the single form
  */
-// TODO: can be refactored into a generic component
 export const EmailField = (): JSX.Element => {
     const {t}: UseTranslationResponse = useTranslation()
