+++ projects/desktop/src/components/common/SingleForm/index.ts
@@ -1,14 +1,17 @@
 export { CenteredForm } from './CenteredForm'
-export { CheckboxContainer } from './checkbox/CheckboxContainer'
-export { CheckboxInput } from './checkbox/CheckboxInput'
-export { CheckboxCheckmark } from './checkbox/CheckboxCheckmark'
 export { FormLink } from './FormLink'
 export { GradientButton } from './GradientButton'
-export { InputBorder } from './input/InputBorder'
-export { InputContainer } from './input/InputContainer'
-export { InputError } from './input/InputError'
-export { InputField } from './input/InputField'
-export { InputWithError } from './input/InputWithError'
 export { BigTitle } from './BigTitle'
-export { Input } from './input/Input'
-export { InputsRow } from './input/InputsRow'
+
+export { InputBorder } from './atoms/input/InputBorder'
+export { InputContainer } from './atoms/input/InputContainer'
+export { InputError } from './atoms/input/InputError'
+export { InputField } from './atoms/input/InputField'
+export { InputWithError } from './atoms/input/InputWithError'
+export { Input } from './atoms/input/Input'
+export { InputsRow } from './atoms/input/InputsRow'
+
+export { CheckboxField } from './fields/CheckboxField'
+export { EmailField } from './fields/EmailField'
+export { PasswordField } from './fields/PasswordField'
+export { TextField } from './fields/TextField'
