+++ projects/desktop/src/components/login/form/index.ts
@@ -1,5 +1,1 @@
-export { Input } from '../../common/singleForm/index'
-export { EmailField } from './EmailField'
-export { PasswordField } from './PasswordField'
-export { CheckboxField } from './CheckboxField'
 export { Form } from './Form'
