+++ projects/desktop/src/components/register/form/Form.tsx
@@ -1,11 +1,9 @@
 import React from 'react'
 import { useTranslation, UseTranslationResponse } from 'react-i18next'
 
-import { CenteredForm, FormLink, InputsRow } from '../../common/singleForm/index'
-import { FirstnameField, LastnameField } from './'
-import { EmailField, PasswordField } from '../../login/form'
+import { CenteredForm, FormLink, InputsRow, CheckboxField, TextField, EmailField, PasswordField } from '../../common/singleForm/index'
+import { required, validateMinimumLength } from '../../../constants/validators'
 import { SubmitButton } from '../styles'
-import { CheckboxField } from './CheckboxField'
 
 interface IOwnProps {
     handleSubmit: () => void,
@@ -23,15 +21,27 @@ export const Form = ({handleSubmit, invalid}: IProps): JSX.Element => {
     return (
         <CenteredForm onSubmit={handleSubmit}>
             <InputsRow>
-                <FirstnameField />
-                <LastnameField />
+                <TextField
+                    name='firstName'
+                    label='register.form.firstname'
+                    validators={[required, validateMinimumLength(2)]}
+                />
+                <TextField
+                    name='lastName'
+                    label='register.form.lastname'
+                    validators={[required, validateMinimumLength(2)]}
+                />
             </InputsRow>
             <InputsRow>
                 <EmailField />
                 <PasswordField />
             </InputsRow>
 
-            <CheckboxField invalid={invalid.termsOfUse} />
+            <CheckboxField
+                name='termsOfUse'
+                textTranslationKey='register.form.termsOfUse'
+                invalid={invalid.termsOfUse}
+            />
 
             <SubmitButton type='submit'>{t('register.form.register')}</SubmitButton>
             <FormLink to='/'>{t('register.form.alreadyRegistered')}</FormLink>
