+++ projects/desktop/src/components/register/form/index.ts
@@ -1,3 +1,1 @@
 export { Form } from './Form'
-export { FirstnameField } from './FirstnameField'
-export { LastnameField } from './LastnameField'
